requirements.txt
streamlit
matplotlib
pandas
seaborn
streamlit_chat
transformers